+++ sound-mind-studio/tests/test_mind_waves_panel.h
@@ -21,4 +21,7 @@ private slots:
     void setMindWavesDropsSelectionWhenTheIdIsGone();
     void clearSelectionDisablesBothEditorsAndEmits();
     void contentIsInAResizableScrollAreaSoThePanelCanShrinkBelowItsFullHeight();
+    void freshPanelHasPreviewOff();
+    void clickingThePreviewButtonEmitsPreviewToggled();
+    void setPreviewEnabledChangesTheButtonWithoutEmitting();
 };